+++ core/src/main/kotlin/org/pastalab/fray/core/observers/AntithesisErrorReporter.kt
@@ -1,12 +1,15 @@
 package org.pastalab.fray.core.observers
 
 import com.antithesis.sdk.Assert
+import com.fasterxml.jackson.databind.ObjectMapper
 import org.pastalab.fray.rmi.TestStatusObserver
 
 class AntithesisErrorReporter : TestStatusObserver {
   override fun onReportError(throwable: Throwable) {
-    Assert.unreachable(
-        "An error occurred during the test execution: ${throwable.message}, stack trace: ${throwable.stackTraceToString()}",
-        null)
+    val objectMapper = ObjectMapper()
+    val node = objectMapper.createObjectNode()
+    node.put("message", throwable.message)
+    node.put("stackTrace", throwable.stackTraceToString())
+    Assert.unreachable("An error occurred during the Fray test", node)
   }
 }